+++ src/components/mainContent/MainContent.tsx
@@ -2,7 +2,6 @@ import './mainContent.styles.css';
 import { useState, useEffect, useContext } from 'react';
 import { Outlet, useNavigate, useSearchParams } from 'react-router-dom';
 
-// Components
 import { Loader } from '../loader/Loader';
 import { Search } from '../search/Search';
 import { PokemonsList } from '../pokemonsList/PokemonsList';
@@ -10,19 +9,10 @@ import { ErrorButton } from '../errorButton/ErrorButton';
 import { Paginator } from '../paginator/Paginator';
 import { PageLimit } from '../pageLimit/PageLimit';
 
-// Utils
-import pokemonApi from '../../api/apiClient';
 import { loadFromLocalStorage } from '../../localStorage/localStorage';
-import { capitalize } from '../../utils/utils';
-import { isConvertibleToInt } from '../../utils/utils';
-import { RightSideContext } from '../../context/context';
-import { DEFAULT_PAGE, INTIAL_ITEM_COUNT } from '../../constants/constants';
-
-interface PokemonItem {
-  name: string;
-  description: string;
-  image: string;
-}
+import { RightSideContext } from '../../context/rightSideContext';
+import { DEFAULT_PAGE } from '../../constants/constants';
+import { PokemonDataContext } from '../../context/pokemonDataContext';
 
 export interface RequestErrors {
   pokemonListRequestError: boolean;
@@ -31,29 +21,20 @@ export interface RequestErrors {
 
 export const MainContent = () => {
   const navigate = useNavigate();
+  const { pokemons, loading, requestErrors, getPokemons, getPokemon, totalItems } = useContext(PokemonDataContext);
+  const context = useContext(RightSideContext);
 
   const [searchParams, setSearchParams] = useSearchParams();
 
   const [searchTerm, setSearchTerm] = useState<string>('');
-  const [pokemons, setPokemons] = useState<PokemonItem[]>([]);
-
-  const [requestErrors, setRequestErrors] = useState<RequestErrors>({
-    pokemonListRequestError: false,
-    pokemonRequestError: false,
-  });
-
   const [causeRenderError, setCauseRenderError] = useState<boolean>(false);
-
-  const [loading, setLoading] = useState<boolean>(false);
 
   const initialPage = parseInt(searchParams.get('page') || '1', 10);
   const initialLimit = parseInt(searchParams.get('limit') || '10', 10);
 
-  const [totalItems, setTotalItems] = useState<number>(INTIAL_ITEM_COUNT);
   const [page, setPage] = useState<number>(initialPage);
   const [limit, setLimit] = useState<number>(initialLimit);
 
-  const context = useContext(RightSideContext);
   const { rightSide, hideRightSide, showRightSide } = context;
 
   useEffect(() => {
@@ -66,99 +47,17 @@ export const MainContent = () => {
     }
 
     const term = loadFromLocalStorage<string>('searchTerm');
-    term && term.length ? getPokemon(term) : getPokemons();
+    term && term.length ? getPokemon(term) : getPokemons(page, limit);
   }, []);
 
   useEffect(() => {
     setSearchParams({ page: page.toString(), limit: limit.toString() });
-    getPokemons();
+    getPokemons(page, limit);
   }, [page, limit]);
-
-  const getPokemons = async (): Promise<void> => {
-    setLoading(true);
-    setRequestErrors({ pokemonListRequestError: false, pokemonRequestError: false });
-
-    const currentOffset = (page - 1) * limit;
-
-    try {
-      const data = await pokemonApi.listPokemons(currentOffset, limit);
-      if (totalItems === INTIAL_ITEM_COUNT) {
-        setTotalItems(data.count);
-      }
-
-      if (data?.results) {
-        const pokemonsData = await fetchAllPokemons(data.results);
-        setPokemons(pokemonsData);
-      } else {
-        throw new Error('No pokemons are found');
-      }
-    } catch {
-      setRequestErrors((prevErrors) => ({
-        ...prevErrors,
-        pokemonListError: true,
-      }));
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const fetchAllPokemons = async (results: { name: string }[]): Promise<PokemonItem[]> => {
-    return Promise.all(
-      results.map(async ({ name }) => {
-        const pokemonDetails = await pokemonApi.getPokemonByName(name);
-        const imageUrl = pokemonDetails.sprites.other?.['official-artwork']['front_default'];
-        return {
-          name,
-          description: `This is a great Pokemon with name ${capitalize(name)} 👻`,
-          image: imageUrl || `default-image-path/${name}.png`,
-        };
-      }),
-    );
-  };
-
-  const getPokemon = async (term: string): Promise<void> => {
-    setLoading(true);
-
-    setRequestErrors({ pokemonListRequestError: false, pokemonRequestError: false });
-
-    if (isConvertibleToInt(term)) {
-      setRequestErrors((prevErrors) => ({
-        ...prevErrors,
-        pokemonRequestError: true,
-      }));
-      return setLoading(false);
-    }
-
-    const nameToSearch = term.toLocaleLowerCase().trim();
-
-    try {
-      const pokemonFromClient = await pokemonApi.getPokemonByName(nameToSearch);
-      const pokemonImage = pokemonFromClient.sprites.other?.['official-artwork']['front_default'];
-
-      if (pokemonImage) {
-        setPokemons([
-          {
-            name: nameToSearch,
-            description: `This is a great Pokemon with name ${capitalize(term)} 👻`,
-            image: pokemonImage,
-          },
-        ]);
-      } else {
-        throw new Error('No pokemon image is found');
-      }
-    } catch {
-      setRequestErrors((prevErrors) => ({
-        ...prevErrors,
-        pokemonRequestError: true,
-      }));
-    } finally {
-      setLoading(false);
-    }
-  };
 
   const onInputChange = async (term: string): Promise<void> => {
     setSearchTerm(term);
-    term.length ? getPokemon(term) : getPokemons();
+    term.length ? getPokemon(term) : getPokemons(page, limit);
   };
 
   const triggerError = (): void => {
